+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # OpenKit Java Changelog
 
-## [Unreleased](https://github.com/Dynatrace/openkit-java/compare/v3.2.0...HEAD)
+## [Unreleased](https://github.com/Dynatrace/openkit-java/compare/v3.3.0...HEAD)
+
+## 3.3.0 [Release date: 2024-08-02]
+[GitHub Releases](https://github.com/Dynatrace/openkit-java/releases/tag/v3.3.0)
 
 ### Added
 - `WebRequestTracer.setBytesSent(long bytes)` to increase the size range